requirements.txt
transparentai>=0.2.1
click>=7.1.2
Flask>=1.1.2
Flask-Babel>=1.0.0
Flask-SQLAlchemy>=2.4.3
Flask-WTF>=0.14.3
gunicorn>=20.0.4
pandas-profiling>=2.8.0